Seperate external methods to services folder and create a Sessions controller
