+++ app/controllers/application_controller.rb
@@ -3,31 +3,12 @@ class ApplicationController < ActionController::API
 
   before_action :authorized
 
-  def encode_token(payload)
-    JWT.encode(payload, 's3cr3t')
-  end
-
-  def auth_header
-    # { Authorization: 'Bearer <token>' }
-    request.headers['Authorization']
-  end
-
-  def decoded_token
-    return unless auth_header
-
-    token = auth_header.split(' ')[1]
-    # header: { 'Authorization': 'Bearer <token>' }
-    begin
-      JWT.decode(token, 's3cr3t', true, algorithm: 'HS256')
-    rescue JWT::DecodeError
-      nil
-    end
-  end
-
   def logged_in_user
-    return unless decoded_token
+    helper = Helper.new(request)
 
-    user_id = decoded_token[0]['user_id']
+    return unless helper.decoded_token
+
+    user_id = helper.decoded_token[0]['user_id']
     @user = User.find_by(id: user_id)
   end
