+++ app/controllers/sessions_controller.rb
@@ -1,20 +1,15 @@
-class UsersController < ApplicationController
-  # Sign up
+class SessionsController < ApplicationController
+  before_action :authorized, except: [:create]
+  # Login
   def create
-    @user = User.create(user_params)
+    @user = User.find_by(email: params[:email])
 
-    if @user.valid?
+    if @user&.authenticate(params[:password])
       helper = Helper.new
       token = helper.encode_token({ user_id: @user.id })
       render json: { user: @user, token: token }
     else
       render json: { error: 'Invalid email or password' }
     end
-  end
-
-  private
-
-  def user_params
-    params.permit(:name, :email, :password)
   end
 end
