+++ app/services/helper.rb
@@ -1,0 +1,26 @@
+class Helper
+  def initialize(request = nil)
+    @request = request
+  end
+
+  def encode_token(payload)
+    JWT.encode(payload, 's3cr3t')
+  end
+
+  def auth_header
+    # { Authorization: 'Bearer <token>' }
+    @request.headers['Authorization']
+  end
+
+  def decoded_token
+    return unless auth_header
+
+    token = auth_header.split(' ')[1]
+    # header: { 'Authorization': 'Bearer <token>' }
+    begin
+      JWT.decode(token, 's3cr3t', true, algorithm: 'HS256')
+    rescue JWT::DecodeError
+      nil
+    end
+  end
+end
