+++ spec/model/user_spec.rb
@@ -10,4 +10,5 @@ RSpec.describe User, type: :model do
   # ensure columns title and created_by are present before saving
   it { should validate_presence_of(:name) }
   it { should validate_presence_of(:email) }
+  it { should validate_presence_of(:password) }
 end
