+++ spec/requests/sessions_request_spec.rb
@@ -1,0 +1,4 @@
+require 'rails_helper'
+
+RSpec.describe 'Sessions', type: :request do
+end
